fix: move ExitResult to error.rs
